+++ packages/expo-sqlite/android/src/main/java/expo/modules/sqlite/SQLiteModule.kt
@@ -8,8 +8,6 @@ import androidx.core.net.toUri
 import androidx.core.os.bundleOf
 import expo.modules.kotlin.exception.Exceptions
 import expo.modules.kotlin.jni.ArrayBuffer
-import expo.modules.kotlin.jni.JavaScriptArrayBuffer
-import expo.modules.kotlin.jni.NativeArrayBuffer
 import expo.modules.kotlin.modules.Module
 import expo.modules.kotlin.modules.ModuleDefinition
 import kotlinx.coroutines.CoroutineScope
@@ -219,10 +217,10 @@ class SQLiteModule : Module() {
         return@Constructor NativeStatement()
       }
 
-      AsyncFunction("runAsync") { statement: NativeStatement, database: NativeDatabase, bindParams: Map<String, Any?>, bindBlobParams: Map<String, NativeArrayBuffer>, shouldPassAsArray: Boolean ->
+      AsyncFunction("runAsync") { statement: NativeStatement, database: NativeDatabase, bindParams: Map<String, Any?>, bindBlobParams: Map<String, ArrayBuffer>, shouldPassAsArray: Boolean ->
         return@AsyncFunction run(statement, database, bindParams, bindBlobParams, shouldPassAsArray)
       }.runOnQueue(moduleCoroutineScope)
-      Function("runSync") { statement: NativeStatement, database: NativeDatabase, bindParams: Map<String, Any?>, bindBlobParams: Map<String, JavaScriptArrayBuffer>, shouldPassAsArray: Boolean ->
+      Function("runSync") { statement: NativeStatement, database: NativeDatabase, bindParams: Map<String, Any?>, bindBlobParams: Map<String, ArrayBuffer>, shouldPassAsArray: Boolean ->
         return@Function run(statement, database, bindParams, bindBlobParams, shouldPassAsArray)
       }
 
@@ -308,17 +306,17 @@ class SQLiteModule : Module() {
         return@Function sessionCreateInvertedChangeset(database, session)
       }
 
-      AsyncFunction("applyChangesetAsync") { session: NativeSession, database: NativeDatabase, changeset: NativeArrayBuffer ->
+      AsyncFunction("applyChangesetAsync") { session: NativeSession, database: NativeDatabase, changeset: ArrayBuffer ->
         sessionApplyChangeset(database, session, changeset)
       }.runOnQueue(moduleCoroutineScope)
-      Function("applyChangesetSync") { session: NativeSession, database: NativeDatabase, changeset: JavaScriptArrayBuffer ->
+      Function("applyChangesetSync") { session: NativeSession, database: NativeDatabase, changeset: ArrayBuffer ->
         sessionApplyChangeset(database, session, changeset)
       }
 
-      AsyncFunction("invertChangesetAsync") { session: NativeSession, database: NativeDatabase, changeset: NativeArrayBuffer ->
+      AsyncFunction("invertChangesetAsync") { session: NativeSession, database: NativeDatabase, changeset: ArrayBuffer ->
         return@AsyncFunction sessionInvertChangeset(database, session, changeset)
       }.runOnQueue(moduleCoroutineScope)
-      Function("invertChangesetSync") { session: NativeSession, database: NativeDatabase, changeset: JavaScriptArrayBuffer ->
+      Function("invertChangesetSync") { session: NativeSession, database: NativeDatabase, changeset: ArrayBuffer ->
         return@Function sessionInvertChangeset(database, session, changeset)
       }
     }
@@ -646,21 +644,21 @@ class SQLiteModule : Module() {
   }
 
   @Throws(AccessClosedResourceException::class, SQLiteErrorException::class)
-  private fun sessionCreateChangeset(database: NativeDatabase, session: NativeSession): NativeArrayBuffer {
+  private fun sessionCreateChangeset(database: NativeDatabase, session: NativeSession): ArrayBuffer {
     maybeThrowForClosedDatabase(database)
     val byteBuffer = session.ref.sqlite3session_changeset()
       ?: throw SQLiteErrorException(database.ref.convertSqlLiteErrorToString())
 
-    return NativeArrayBuffer(byteBuffer)
+    return ArrayBuffer(byteBuffer)
   }
 
   @Throws(AccessClosedResourceException::class, SQLiteErrorException::class)
-  private fun sessionCreateInvertedChangeset(database: NativeDatabase, session: NativeSession): NativeArrayBuffer {
+  private fun sessionCreateInvertedChangeset(database: NativeDatabase, session: NativeSession): ArrayBuffer {
     maybeThrowForClosedDatabase(database)
     val byteBuffer = session.ref.sqlite3session_changeset_inverted()
       ?: throw SQLiteErrorException(database.ref.convertSqlLiteErrorToString())
 
-    return NativeArrayBuffer(byteBuffer)
+    return ArrayBuffer(byteBuffer)
   }
 
   @Throws(AccessClosedResourceException::class, SQLiteErrorException::class)
@@ -672,12 +670,12 @@ class SQLiteModule : Module() {
   }
 
   @Throws(AccessClosedResourceException::class, SQLiteErrorException::class)
-  private fun sessionInvertChangeset(database: NativeDatabase, session: NativeSession, changeset: ArrayBuffer): NativeArrayBuffer {
+  private fun sessionInvertChangeset(database: NativeDatabase, session: NativeSession, changeset: ArrayBuffer): ArrayBuffer {
     maybeThrowForClosedDatabase(database)
     val byteBuffer = session.ref.sqlite3changeset_invert(changeset.toDirectBuffer())
       ?: throw SQLiteErrorException(database.ref.convertSqlLiteErrorToString())
 
-    return NativeArrayBuffer(byteBuffer)
+    return ArrayBuffer(byteBuffer)
   }
 
   // endregion